refactor: use ui/swiper
wrapping Swiper with SwiperContainer
